move marksweep::metadata to mallocspace
