+++ src/plan/marksweep/mod.rs
@@ -1,6 +1,5 @@
 mod gc_work;
 mod global;
-pub mod metadata;
 pub mod mutator;
 
 pub use self::global::MarkSweep;
