+++ src/policy/mallocspace/mod.rs
@@ -1,0 +1,4 @@
+mod global;
+pub mod metadata;
+
+pub use global::*;
